split shift calendar renderer into different classes
