+++ includes/engelsystem_provider.php
@@ -31,6 +31,7 @@ require_once realpath(__DIR__ . '/../includes/view/Questions_view.php');
 require_once realpath(__DIR__ . '/../includes/view/Rooms_view.php');
 require_once realpath(__DIR__ . '/../includes/view/ShiftCalendarLane.php');
 require_once realpath(__DIR__ . '/../includes/view/ShiftCalendarRenderer.php');
+require_once realpath(__DIR__ . '/../includes/view/ShiftCalendarShiftRenderer.php');
 require_once realpath(__DIR__ . '/../includes/view/ShiftsFilterRenderer.php');
 require_once realpath(__DIR__ . '/../includes/view/Shifts_view.php');
 require_once realpath(__DIR__ . '/../includes/view/ShiftEntry_view.php');
